update profile details

+++ app/Http/Controllers/API/UserController.php
@@ -16,7 +16,7 @@ class UserController extends Controller
      */
      public function __construct()
      {
-        //  $this->middleware('auth');
+        $this->middleware('auth:api');
      }
 
     /**
@@ -26,6 +26,10 @@ class UserController extends Controller
      */
     public function index()
     {
+        if(\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            return User::latest()->paginate(10);
+        }
+        // $this->authorize('isAdmin');
         return User::latest()->paginate(10);
     }
 
@@ -67,6 +71,60 @@ class UserController extends Controller
         //
     }
 
+    public function profile()
+    {
+        return auth('api')->user();
+        // $user = User::findOrFail($id);
+        // return $user;
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+        // $user = User::findOrFail($id);
+
+        $this->validate($request, [
+            'name' => 'required|string|max:
+            255',
+            'email' => 'required|email|string|max:255|unique:users,email,'.$user->id,
+            'password' => 'sometimes|string|min:6
+            '
+        ]);
+
+        if(!empty($request->photo)) {
+            $currentPhoto = $user->photo;
+            if($request->photo != $currentPhoto) {
+                // $name = time().'.'.'jpg';
+                $name = time().'.'. explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
+                // dd($name);
+                \Image::make($request->photo)->save(public_path('img/profile/').$name);
+
+                $request->merge(['photo' => $name]);
+
+                $userPhoto = public_path('img/profile/').$currentPhoto;
+                if (file_exists($userPhoto)) {
+                    @unlink($userPhoto);
+                }
+
+            }
+        }
+
+         if(!empty($request->password)) {
+             $request->merge(['password' => Hash::make($request['password'])]);
+         }
+
+         $user->update($request->all());
+        return ['Message' => 'Success'];
+    }
+
     /**
      * Update the specified resource in storage.
      *
@@ -97,8 +155,26 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-       $user = User::findOrFail($id);
+        $this->authorize('isAdmin');
 
-       $user->delete();
+        $user = User::findOrFail($id);
+
+        $user->delete();
+    }
+
+    public function findUser() {
+
+        if($search = \Request::get('q')) {
+            $user = User::where(function($query) use ($search) {
+                $query->where('name', 'LIKE', "%$search%")
+                        ->orWhere('email', 'LIKE', "%$search%")
+                        ->orWhere('role', 'LIKE', "%$search%");
+            })->paginate(5);
+
+        } else {
+            $user = User::latest()->paginate(10);
+        }
+
+        return $user;   
     }
 }
